requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
aiohttp==3.9.1
python-multipart==0.0.6
twilio==8.10.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
analytics-python==1.4.0
apscheduler==3.10.4
alembic==1.13.1
sendgrid==6.10.0
python-socketio==5.10.0
amadeus==8.1.0
redis==5.0.1
celery==5.3.6
prometheus-client==0.19.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
